# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Rest para SocialMeli (followers, posts, promoçōes)",
        "title": "SocialMeli API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/products/followed/latest/{userId}": {
            "get": {
                "description": "Retorna publicações dos vendedores seguidos nas últimas semanas (default=2). Suporta ordenação por data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feed"
                ],
                "summary": "Feed de publicações de vendedores seguidos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "date_asc",
                            "date_desc"
                        ],
                        "type": "string",
                        "description": "Ordenação",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Quantidade de semanas (default 2)",
                        "name": "weeks",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.FollowedPostsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/followed/{userId}/list": {
            "get": {
                "description": "Retorna publicações dos vendedores seguidos nas últimas semanas (default=2). Suporta ordenação por data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feed"
                ],
                "summary": "Feed de publicações de vendedores seguidos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "date_asc",
                            "date_desc"
                        ],
                        "type": "string",
                        "description": "Ordenação",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Quantidade de semanas (default 2)",
                        "name": "weeks",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.FollowedPostsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/sellers/{sellerId}/promotions/count": {
            "get": {
                "description": "Retorna a quantidade de publicações promocionais de um vendedor",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promotions"
                ],
                "summary": "Contagem de promoções por vendedor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do vendedor",
                        "name": "sellerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SellerPromotionsCountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Retorna todos os usuários",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Listar usuários",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.UsersListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um novo usuário (seller ou buyer)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Criar usuário",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "description": "Retorna um usuário pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Detalhar usuário",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza nome e/ou flag de vendedor",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Atualizar usuário",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos atualizáveis",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}/follow/{sellerId}": {
            "post": {
                "description": "Cria relacionamento de follow entre usuário e vendedor",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Follow"
                ],
                "summary": "Seguir vendedor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário (follower)",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID do vendedor (seller)",
                        "name": "sellerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.FollowResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove relacionamento de follow existente",
                "tags": [
                    "Follow"
                ],
                "summary": "Deixar de seguir vendedor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário (follower)",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID do vendedor (seller)",
                        "name": "sellerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}/followed/list": {
            "get": {
                "description": "Retorna a lista de vendedores seguidos por um usuário. Permite ordenação por nome.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Following"
                ],
                "summary": "Lista de vendedores seguidos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "name_asc",
                            "name_desc"
                        ],
                        "type": "string",
                        "description": "Ordenação",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.FollowingListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}/followers/count": {
            "get": {
                "description": "Retorna a quantidade de seguidores de um vendedor",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Followers"
                ],
                "summary": "Contagem de seguidores",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário (vendedor)",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.FollowersCountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_user_infrastructure_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.CreateUserRequest": {
            "type": "object",
            "properties": {
                "is_seller": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "api.FollowResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "api.FollowedPostsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.FeedPostResponse"
                    }
                },
                "user_id": {
                    "type": "integer"
                },
                "weeks": {
                    "type": "integer"
                }
            }
        },
        "api.FollowersCountResponse": {
            "type": "object",
            "properties": {
                "followers_count": {
                    "type": "integer"
                }
            }
        },
...
        "api.FollowingListResponse": {
            "type": "object",
            "properties": {
                "following": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.UserResponse"
                    }
                }
            }
        },
...
        "api.SellerPromotionsCountResponse": {
            "type": "object",
            "properties": {
                "promo_count": {
                    "type": "integer"
                },
                "seller_id": {
                    "type": "integer"
                }
            }
        },
        "api.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_seller": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "api.UsersListResponse": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.UserResponse"
                    }
                }
            }
        },
...
        "internal_user_infrastructure_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}